Fix acceptance tests paths
