+++ test/acceptance/inquirer/CliRunner.js
