+++ test/acceptance/inquirer/autocomplete.spec.js
