+++ test/acceptance/inquirer/custom-quit.spec.js
