+++ test/acceptance/inquirer/exit-logs-mode.spec.js
