+++ test/acceptance/inquirer/fixtures/custom-quit-method.js
@@ -9,7 +9,7 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-const { Inquirer } = require("../../../../lib/cli/Inquirer");
+const { Inquirer } = require("../../../../src/Inquirer");
 
 const questions = {
   main: {
@@ -20,10 +20,6 @@ const questions = {
       {
         name: "Option 1",
         value: "option1"
-      },
-      {
-        name: "Display logs",
-        value: "logs"
       }
     ]
   }
@@ -31,7 +27,10 @@ const questions = {
 
 const MyCli = class MyCli {
   constructor() {
-    this._cli = new Inquirer(questions, this.header.bind(this));
+    this._cli = new Inquirer(questions, this.header.bind(this), {
+      name: "Custom Quit",
+      action: this.customExitSelected.bind(this)
+    });
     this._selectedOption = "None";
   }
 
@@ -39,21 +38,14 @@ const MyCli = class MyCli {
     return [`Selected option: ${this._selectedOption}`];
   }
 
-  async displayLogs() {
-    await this._cli.logsMode(() => {
-      console.log("This is a foo log");
-      setTimeout(() => {
-        this._cli.exitLogsMode();
-      }, 500);
-    });
+  async customExitSelected() {
+    this._selectedOption = "Quit";
+    await this.displayMainMenu();
   }
 
   async displayMainMenu() {
     this._cli.clearScreen();
     this._selectedOption = await this._cli.inquire("main");
-    if (this._selectedOption === "logs") {
-      await this.displayLogs();
-    }
     await this.displayMainMenu();
   }
 };
