+++ test/acceptance/inquirer/fixtures/exit-logs-mode.js
@@ -9,7 +9,7 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-const { Inquirer } = require("../../../../lib/cli/Inquirer");
+const { Inquirer } = require("../../../../src/Inquirer");
 
 const questions = {
   main: {
@@ -18,12 +18,12 @@ const questions = {
     name: "value",
     choices: [
       {
-        name: "Main option 1",
+        name: "Option 1",
         value: "option1"
       },
       {
-        name: "Main option 2",
-        value: "option2"
+        name: "Display logs",
+        value: "logs"
       }
     ]
   }
@@ -39,9 +39,21 @@ const MyCli = class MyCli {
     return [`Selected option: ${this._selectedOption}`];
   }
 
+  async displayLogs() {
+    await this._cli.logsMode(() => {
+      console.log("This is a foo log");
+      setTimeout(() => {
+        this._cli.exitLogsMode();
+      }, 500);
+    });
+  }
+
   async displayMainMenu() {
     this._cli.clearScreen();
     this._selectedOption = await this._cli.inquire("main");
+    if (this._selectedOption === "logs") {
+      await this.displayLogs();
+    }
     await this.displayMainMenu();
   }
 };
