+++ test/acceptance/inquirer/fixtures/readme-example.js
@@ -9,7 +9,7 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-const { Inquirer } = require("../../../../lib/cli/Inquirer");
+const { Inquirer } = require("../../../../src/Inquirer");
 
 const questions = {
   main: {
@@ -18,8 +18,12 @@ const questions = {
     name: "value",
     choices: [
       {
-        name: "Option 1",
+        name: "Main option 1",
         value: "option1"
+      },
+      {
+        name: "Main option 2",
+        value: "option2"
       }
     ]
   }
@@ -27,20 +31,12 @@ const questions = {
 
 const MyCli = class MyCli {
   constructor() {
-    this._cli = new Inquirer(questions, this.header.bind(this), {
-      name: "Custom Quit",
-      action: this.customExitSelected.bind(this)
-    });
+    this._cli = new Inquirer(questions, this.header.bind(this));
     this._selectedOption = "None";
   }
 
   header() {
     return [`Selected option: ${this._selectedOption}`];
-  }
-
-  async customExitSelected() {
-    this._selectedOption = "Quit";
-    await this.displayMainMenu();
   }
 
   async displayMainMenu() {
