+++ test/acceptance/inquirer/logs-mode.spec.js
