+++ test/acceptance/inquirer/readme-example.spec.js
