+++ test/acceptance/inquirer/remove-listeners.spec.js
