+++ test/acceptance/main/cli-disabled-arguments.spec.js
@@ -9,10 +9,9 @@ Unless required by applicable law or agreed to in writing, software distributed 
 */
 
 const path = require("path");
-const { CliRunner, request, wait, TimeCounter } = require("./utils");
+const { CliRunner, request, wait, TimeCounter, BINARY_PATH } = require("./utils");
 
 describe("command line arguments with cli disabled", () => {
-  const binaryPath = "../../../../bin/mocks-server";
   const cwdPath = path.resolve(__dirname, "fixtures");
   let cli;
 
@@ -22,7 +21,7 @@ describe("command line arguments with cli disabled", () => {
 
   describe("interactive cli", () => {
     it("should not be started", async () => {
-      cli = new CliRunner([binaryPath, "--behaviors=web-tutorial", "--cli=false"], {
+      cli = new CliRunner([BINARY_PATH, "--behaviors=web-tutorial", "--cli=false"], {
         cwd: cwdPath
       });
       await wait(3000);
@@ -32,7 +31,7 @@ describe("command line arguments with cli disabled", () => {
 
   describe("behaviors option", () => {
     it("should set mocks folder", async () => {
-      cli = new CliRunner([binaryPath, "--behaviors=web-tutorial", "--cli=false"], {
+      cli = new CliRunner([BINARY_PATH, "--behaviors=web-tutorial", "--cli=false"], {
         cwd: cwdPath
       });
       await wait();
@@ -47,7 +46,7 @@ describe("command line arguments with cli disabled", () => {
   describe("behavior option", () => {
     describe("when not provided", () => {
       it("should set as current behavior the first one found", async () => {
-        cli = new CliRunner([binaryPath, "--behaviors=web-tutorial", "--cli=false"], {
+        cli = new CliRunner([BINARY_PATH, "--behaviors=web-tutorial", "--cli=false"], {
           cwd: cwdPath
         });
         await wait();
@@ -59,7 +58,7 @@ describe("command line arguments with cli disabled", () => {
     describe("when provided and exists", () => {
       it("should set current behavior", async () => {
         cli = new CliRunner(
-          [binaryPath, "--behaviors=web-tutorial", "--behavior=dynamic", "--cli=false"],
+          [BINARY_PATH, "--behaviors=web-tutorial", "--behavior=dynamic", "--cli=false"],
           {
             cwd: cwdPath
           }
@@ -73,7 +72,7 @@ describe("command line arguments with cli disabled", () => {
     describe("when provided and does not exist", () => {
       it("should print a warning", async () => {
         cli = new CliRunner(
-          [binaryPath, "--behaviors=web-tutorial", "--behavior=foo", "--cli=false"],
+          [BINARY_PATH, "--behaviors=web-tutorial", "--behavior=foo", "--cli=false"],
           {
             cwd: cwdPath
           }
@@ -84,7 +83,7 @@ describe("command line arguments with cli disabled", () => {
 
       it("should set as current behavior the first one found", async () => {
         cli = new CliRunner(
-          [binaryPath, "--behaviors=web-tutorial", "--behavior=foo", "--cli=false"],
+          [BINARY_PATH, "--behaviors=web-tutorial", "--behavior=foo", "--cli=false"],
           {
             cwd: cwdPath
           }
@@ -98,7 +97,7 @@ describe("command line arguments with cli disabled", () => {
 
   describe("features option", () => {
     it("should set mocks folder", async () => {
-      cli = new CliRunner([binaryPath, "--features=web-tutorial", "--cli=false"], {
+      cli = new CliRunner([BINARY_PATH, "--features=web-tutorial", "--cli=false"], {
         cwd: cwdPath
       });
       await wait();
@@ -113,7 +112,7 @@ describe("command line arguments with cli disabled", () => {
   describe("feature option", () => {
     describe("when not provided", () => {
       it("should set as current behavior the first one found", async () => {
-        cli = new CliRunner([binaryPath, "--features=web-tutorial", "--cli=false"], {
+        cli = new CliRunner([BINARY_PATH, "--features=web-tutorial", "--cli=false"], {
           cwd: cwdPath
         });
         await wait();
@@ -125,7 +124,7 @@ describe("command line arguments with cli disabled", () => {
     describe("when provided and exists", () => {
       it("should set current behavior", async () => {
         cli = new CliRunner(
-          [binaryPath, "--features=web-tutorial", "--feature=dynamic", "--cli=false"],
+          [BINARY_PATH, "--features=web-tutorial", "--feature=dynamic", "--cli=false"],
           {
             cwd: cwdPath
           }
@@ -139,7 +138,7 @@ describe("command line arguments with cli disabled", () => {
     describe("when provided and does not exist", () => {
       it("should print a warning", async () => {
         cli = new CliRunner(
-          [binaryPath, "--features=web-tutorial", "--feature=foo", "--cli=false"],
+          [BINARY_PATH, "--features=web-tutorial", "--feature=foo", "--cli=false"],
           {
             cwd: cwdPath
           }
@@ -150,7 +149,7 @@ describe("command line arguments with cli disabled", () => {
 
       it("should set as current behavior the first one found", async () => {
         cli = new CliRunner(
-          [binaryPath, "--features=web-tutorial", "--feature=foo", "--cli=false"],
+          [BINARY_PATH, "--features=web-tutorial", "--feature=foo", "--cli=false"],
           {
             cwd: cwdPath
           }
@@ -166,7 +165,7 @@ describe("command line arguments with cli disabled", () => {
     it("should set delay", async () => {
       expect.assertions(2);
       cli = new CliRunner(
-        [binaryPath, "--behaviors=web-tutorial", "--delay=2000", "--cli=false"],
+        [BINARY_PATH, "--behaviors=web-tutorial", "--delay=2000", "--cli=false"],
         {
           cwd: cwdPath
         }
