--- test/acceptance/main/fixtures/programmatic-cli/start-unexistant-behavior.js
@@ -1,14 +1,0 @@
-const path = require("path");
-const { Cli } = require("../../../../../index");
-
-const cli = new Cli({
-  behaviors: path.resolve(__dirname, "..", "web-tutorial"),
-  behavior: "foo",
-  port: 3100,
-  log: "debug",
-  watch: false
-});
-
-cli.start().catch(err => {
-  console.log("Error starting CLI", err);
-});
