--- test/acceptance/main/fixtures/programmatic-server/start-delay.js
@@ -1,12 +1,0 @@
-const path = require("path");
-const { Server } = require("../../../../../index");
-
-const server = new Server(path.resolve(__dirname, "..", "web-tutorial"), {
-  port: 3100,
-  log: "debug",
-  delay: 2000
-});
-
-server.start().then(() => {
-  console.log("Server started");
-});
