--- test/acceptance/main/fixtures/programmatic-server/start-watch.js
@@ -1,12 +1,0 @@
-const path = require("path");
-const { Server } = require("../../../../../index");
-
-const server = new Server(path.resolve(__dirname, "..", "files-watch"), {
-  port: 3100,
-  log: "debug",
-  watch: true
-});
-
-server.start().then(() => {
-  console.log("Server started");
-});
