--- test/acceptance/main/web-tutorial-admin-api.spec.js
@@ -1,104 +1,0 @@
-/*
-Copyright 2019 Javier Brea
-
-Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
-
-http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
-*/
-
-const { startServer, stopServer, request, changeBehavior, getBehaviors } = require("./utils");
-
-describe("web tutorial", () => {
-  let server;
-
-  beforeAll(async () => {
-    server = await startServer();
-  });
-
-  afterAll(() => {
-    stopServer(server);
-  });
-
-  describe("When started", () => {
-    it("should have 3 behaviors available", async () => {
-      const behaviors = await getBehaviors();
-      expect(behaviors.length).toEqual(3);
-    });
-
-    it("should serve users collection mock under the /api/users path", async () => {
-      const users = await request("/api/users");
-      expect(users).toEqual([
-        { id: 1, name: "John Doe" },
-        { id: 2, name: "Jane Doe" }
-      ]);
-    });
-
-    it("should serve user 1 under the /api/users/1 path", async () => {
-      const users = await request("/api/users/1");
-      expect(users).toEqual({ id: 1, name: "John Doe" });
-    });
-
-    it("should serve user 1 under the /api/users/2 path", async () => {
-      const users = await request("/api/users/2");
-      expect(users).toEqual({ id: 1, name: "John Doe" });
-    });
-  });
-
-  describe('When changing current behavior to "user2"', () => {
-    beforeAll(async () => {
-      await changeBehavior("user2");
-    });
-
-    it("should serve users collection mock under the /api/users path", async () => {
-      const users = await request("/api/users");
-      expect(users).toEqual([
-        { id: 1, name: "John Doe" },
-        { id: 2, name: "Jane Doe" }
-      ]);
-    });
-
-    it("should serve user 2 under the /api/users/1 path", async () => {
-      const users = await request("/api/users/1");
-      expect(users).toEqual({ id: 2, name: "Jane Doe" });
-    });
-
-    it("should serve user 2 under the /api/users/2 path", async () => {
-      const users = await request("/api/users/2");
-      expect(users).toEqual({ id: 2, name: "Jane Doe" });
-    });
-  });
-
-  describe('When changing current behavior to "dynamic"', () => {
-    beforeAll(async () => {
-      await changeBehavior("dynamic");
-    });
-
-    it("should serve users collection mock under the /api/users path", async () => {
-      const users = await request("/api/users");
-      expect(users).toEqual([
-        { id: 1, name: "John Doe" },
-        { id: 2, name: "Jane Doe" }
-      ]);
-    });
-
-    it("should serve user 1 under the /api/users/1 path", async () => {
-      const users = await request("/api/users/1");
-      expect(users).toEqual({ id: 1, name: "John Doe" });
-    });
-
-    it("should serve user 2 under the /api/users/2 path", async () => {
-      const users = await request("/api/users/2");
-      expect(users).toEqual({ id: 2, name: "Jane Doe" });
-    });
-
-    it("should return not found for /api/users/3 path", async () => {
-      const usersResponse = await request("/api/users/3", {
-        resolveWithFullResponse: true,
-        simple: false
-      });
-      expect(usersResponse.statusCode).toEqual(404);
-    });
-  });
-});
