+++ test/acceptance/mocks-server
@@ -1,5 +1,10 @@
+#!/usr/bin/env node
+"use strict";
+
+require("./start").start();
+
 /*
-Copyright 2019 Javier Brea
+Copyright 2019 XbyOrange
 
 Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
 
@@ -7,23 +12,3 @@ http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
-
-// /mocks/behaviors.js
-
-const { Behavior } = require("@mocks-server/core");
-
-const { getUsers, getUser, getUser2, getRealUser } = require("./fixtures/users");
-
-const standard = new Behavior([getUsers, getUser]);
-
-// Extends the standard behavior adding "getUser2" fixture.
-const user2 = standard.extend([getUser2]);
-
-// Extends the standard behavior adding "getRealUser" dynamic fixture.
-const dynamic = standard.extend([getRealUser]);
-
-module.exports = {
-  standard,
-  user2,
-  dynamic
-};
